move mnu types handler into separate file
